requirements.txt
streamlit
google-generativeai
openai
youtube_transcript_api
python-dotenv
beautifulsoup4
st-copy-to-clipboard
strip-markdown
yt-dlp
requests
transformers
torch
sentencepiece
opencv-python
Pillow
numpy
pytesseract
tensorflow
keras